pyproject.toml
[tool.poetry]
name = "test-anthropic"
version = "0.1.0"
description = ""
authors = ["oriojas <oscarriojas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.33.0"
anthropic = "^0.25.7"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
